Player Controller Update
Check Content>Jade>BPs for new update

Added Wallrun:

Combined player adjustments from Zach with Wallrun logic. Also added in a new collision profile for runnable walls. This gives us the ability to determine whether or not players can run on specific walls. Useful for traps if we add them.

All of which is done in a separate file.