More Bugfixing
Did the following in this commit:

Fixed bug when falling and pressing sprint button that would cause the player's velocity to be reset. (Culprit was in the reset wall run function, it was only setting movement mode to walking for character movement).

Reduced the power of double jump when the player is not wall clinging.

Fixed a missing texture bug on a single mesh in Jade Assets.