Jade Group Level 2 Update
Updated Jade Group Level 2 with a new path to explore, fog particles, and a few decorations. I would say the level is about 90-95% done. Last few things it needs are decorations and details to fill in the spaces a bit.

Added animation for carrying objects to ABP_DEVVY.

Adjusted Socket Position for carrying objects.

Adjusted current montages to allow for animation updates in the future (Added a slot)

Created Light and Dense fog particle systems